refactor(UI): :recycle: Group all user api functions in one file
