--- meiki_ui/src/lib/api/authStatus.ts
@@ -1,5 +1,0 @@
-import { makeRequest } from "./request"
-
-export async function authStatus() {
-    await makeRequest("/authStatus", "GET")
-}
