--- meiki_ui/src/lib/api/createAccount.ts
@@ -1,10 +1,0 @@
-import { makeRequest } from "./request"
-
-export async function createAccount(username: string, password: string) {
-    const body = {
-        username,
-        password,
-    }
-
-    await makeRequest("/create", "POST", body)
-}
