--- meiki_ui/src/lib/api/delete.ts
@@ -1,10 +1,0 @@
-import { makeRequest } from "./request"
-
-export async function del(username: string, password: string) {
-    const body = {
-        username,
-        password,
-    }
-
-    await makeRequest("/delete", "POST", body)
-}
