--- meiki_ui/src/lib/api/logout.ts
@@ -1,6 +1,0 @@
-import { makeRequest } from "./request"
-
-export async function logout() {
-    await makeRequest("/logout", "POST", {})
-    localStorage.clear()
-}
